+++ src/node/utils.ts
@@ -1,9 +1,12 @@
 import path from 'path'
-import { promises as fs } from 'fs'
+import fs from 'fs-extra'
 import LRUCache from 'lru-cache'
 import { Context } from 'koa'
 import { Readable } from 'stream'
 const getETag = require('etag')
+
+const httpRE = /^https?:\/\//
+export const isExternalUrl = (url: string) => httpRE.test(url)
 
 const imageRE = /\.(png|jpe?g|gif|svg)(\?.*)?$/
 const mediaRE = /\.(mp4|webm|ogg|mp3|wav|flac|aac)(\?.*)?$/