+++ test/test.js
@@ -1,4 +1,4 @@
-const fs = require('fs').promises
+const fs = require('fs-extra')
 const path = require('path')
 const execa = require('execa')
 const puppeteer = require('puppeteer')
@@ -13,18 +13,12 @@ let browser
 jest.setTimeout(100000)
 
 beforeAll(async () => {
-  await fs.rmdir(tempDir, { recursive: true })
-  await fs.mkdir(tempDir)
-  for (const file of await fs.readdir(fixtureDir)) {
-    await fs.copyFile(
-      path.join(__dirname, 'fixtures', file),
-      path.join(tempDir, file)
-    )
-  }
+  await fs.remove(tempDir)
+  await fs.copy(fixtureDir, tempDir)
 })
 
 afterAll(async () => {
-  await fs.rmdir(tempDir, { recursive: true })
+  await fs.remove(tempDir)
   if (browser) await browser.close()
   if (server)
     server.kill('SIGTERM', {
@@ -84,7 +78,7 @@ describe('vite', () => {
     )
 
     const button = await page.$('button')
-    testByPolling('count is 1', () => {
+    await testByPolling('count is 1', () => {
       return button.evaluate((b) => b.textContent)
     })
   })
@@ -106,7 +100,7 @@ describe('vite', () => {
     )
 
     const child = await page.$('.child')
-    testByPolling('rgb(255, 0, 0)', () => {
+    await testByPolling('rgb(255, 0, 0)', () => {
       return child.evaluate((e) => getComputedStyle(e).color)
     })
   })
@@ -115,12 +109,12 @@ describe('vite', () => {
 })
 
 // poll until it updates
-async function testByPolling(expect, poll) {
+async function testByPolling(expectValue, poll) {
   const maxTries = 10
   for (let tries = 0; tries < maxTries; tries++) {
     const actual = await poll()
-    if (actual === expect || tries === maxTries - 1) {
-      expect(actual).toBe(expect)
+    if (actual === expectValue || tries === maxTries - 1) {
+      expect(actual).toBe(expectValue)
     } else {
       await timeout(200)
     }